+++ src/main/java/org/sopt/snappinserver/global/response/code/product/ProductSuccessCode.java
@@ -1,4 +1,4 @@
-package org.sopt.snappinserver.api.v1.product.code;
+package org.sopt.snappinserver.global.response.code.product;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
